requirements.txt
fastapi[standard]==0.112.1
uvicorn == 0.30.6
pydantic == 2.8.2
supabase == 2.7.3
python-dotenv == 1.0.1